Fix math bug on 16-bit processors
When scaling analog input values, cast the state variable to uint32_t
to avoid truncating the multiplication.  This avoids problems on
processors where an int is smaller than 32 bits.

+++ MidiController.ino
@@ -276,7 +276,8 @@ void loop() {
       else if (state > p->Max) {
         state = p->Max;
       }
-      state = 1023 * (state - p->Min) / (p->Max - p->Min);
+      // use a wide enough integer for the intermediate values
+      state = (uint32_t)1023 * (state - p->Min) / (p->Max - p->Min);
       break;
     case Digital:
     case DigitalPullup:
